Code Optimization: Use private/pub(crate) to limit the scope of object access in canopen.
Additionally, it includes:

Removed all warnings during compilation. Some temporarily bypassed with #allow[], which will be reviewed later.
Used Rust's "? operator" for clearer logic and to enforce consideration of the use of Result<T, E>.
Provided debug!/warn!/error! in the prelude.
Prepared a run_coverage script for use in CI/CD to ensure code coverage is not too poor.